database connection refactor
