+++ services/user-service/src/configurations/envVariables.ts
