+++ services/user-service/src/database/pool.ts
@@ -1,6 +1,6 @@
 import { Pool } from 'pg';
 
-export const databasePool = new Pool({
+const databasePool = new Pool({
     user: process.env.DB_USER,
     password: process.env.DB_PASSWORD,
     database: process.env.DB_NAME,
@@ -8,10 +8,7 @@ export const databasePool = new Pool({
     port: parseInt(process.env.DB_PORT || '5432', 10)
 });
 
-databasePool.on('connect', () => console.log('PostgreSQL pool conected'));
+databasePool.on('connect', () => console.log('DB pool connected'));
 databasePool.on('error', (err) => console.error('DB Pool error:', err));
 
-export async function closeDb() {
-    await databasePool.end();
-    console.log('PostgreSQL pool closed');
-}
+export default databasePool;
